moving set signal handler to network utils
